Break out S3ReservationsWriter
